Simplify KZG tests
